HW 21. Поправлен trackTimeMillis на long

+++ app/src/main/java/com/practicum/playlistmaker/search/data/db/entity/FavoritesEntity.kt
@@ -9,7 +9,7 @@ data class FavoritesEntity(
     @PrimaryKey val trackId: String,
     val trackName: String,
     val artistName: String,
-    val trackTimeMillis: Int,
+    val trackTimeMillis: Long,
     val artworkUrl100: String,
     val collectionName: String,
     val releaseDate: String,
